Coverage has been moved to bamops. I hope this isn't contentious
